+++ src-docs/src/routes.js
@@ -153,6 +153,9 @@ import { LoadingExample }
 import { ModalExample }
   from './views/modal/modal_example';
 
+import { MutationObserverExample }
+  from './views/mutation_observer/mutation_observer_example';
+
 import { OutsideClickDetectorExample }
   from './views/outside_click_detector/outside_click_detector_example';
 
@@ -383,6 +386,7 @@ const navigation = [{
     PortalExample,
     ToggleExample,
     UtilityClassesExample,
+    MutationObserverExample,
   ].map(example => createExample(example)),
 }, {
   name: 'Package',