+++ src/components/portal/portal.js
@@ -5,7 +5,14 @@
 
 import { Component } from 'react';
 import PropTypes from 'prop-types';
-import { createPortal } from 'react-dom';
+import { createPortal, findDOMNode } from 'react-dom';
+
+export const insertPositions = {
+  'after': 'afterend',
+  'before': 'beforebegin',
+};
+
+export const INSERT_POSITIONS = Object.keys(insertPositions);
 
 export class EuiPortal extends Component {
   constructor(props) {
@@ -13,15 +20,37 @@ export class EuiPortal extends Component {
 
     const {
       children, // eslint-disable-line no-unused-vars
+      insert,
     } = this.props;
 
     this.portalNode = document.createElement('div');
-    document.body.appendChild(this.portalNode);
+
+    if (insert == null) {
+      // no insertion defined, append to body
+      document.body.appendChild(this.portalNode);
+    } else {
+      // inserting before or after an element
+      findDOMNode(insert.sibling).insertAdjacentElement(
+        insertPositions[insert.position],
+        this.portalNode
+      );
+    }
+  }
+
+  componentDidMount() {
+    this.updatePortalRef();
   }
 
   componentWillUnmount() {
-    document.body.removeChild(this.portalNode);
+    this.portalNode.parentNode.removeChild(this.portalNode);
     this.portalNode = null;
+    this.updatePortalRef();
+  }
+
+  updatePortalRef() {
+    if (this.props.portalRef) {
+      this.props.portalRef(this.portalNode);
+    }
   }
 
   render() {
@@ -34,4 +63,13 @@ export class EuiPortal extends Component {
 
 EuiPortal.propTypes = {
   children: PropTypes.node,
+  /** `{sibling: ReactNode|HTMLElement, position: 'before'|'after'}` */
+  insert: PropTypes.shape({
+    sibling: PropTypes.oneOfType([
+      PropTypes.node,
+      PropTypes.instanceOf(HTMLElement)
+    ]).isRequired,
+    position: PropTypes.oneOf(INSERT_POSITIONS),
+  }),
+  portalRef: PropTypes.func,
 };